+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - FIXED: **[BREAKING]** Export factory function on `L.tileLayer.bing` not `L.TileLayer.bing`
 - IMPROVED: Package with browserify and require dependencies
+- ADDED: `getMetaData` method
 
 ## v2.0.2 - 2015-12-03